Paramaters still hard coded in ImportData.
Both steps to create the index and run the analysis can be run from command line.

param build will create the index in the Directory specified in ImportData
param run will use the index in ImportData and the file and column specified in ImportData. Exports to the file specified in ImportData.